Prototype of many agents example

+++ examples/many_agents/options.rs
@@ -6,10 +6,10 @@ use clap::Parser;
 #[command(author, version, about, long_about = None)]
 pub struct ReplicatorOptions{
 
-    #[arg(short = 'v', long = "log_level", value_enum, default_value = "debug")]
+    #[arg(short = 'v', long = "log_level", value_enum, default_value = "info")]
     pub log_level: LevelFilter,
 
-    #[arg(short = 'a', long = "log_level_amfi", value_enum, default_value = "OFF")]
+    #[arg(short = 'a', long = "log_level_amfi", value_enum, default_value = "debug")]
     pub log_level_amfi: LevelFilter,
 
     #[arg(short = 'o', long = "logfile")]
@@ -30,17 +30,17 @@ pub struct ReplicatorOptions{
     #[arg(short = 'n', long = "rounds", default_value = "32")]
     pub number_of_rounds: usize,
 
-    #[arg(short = 'h', long = "hawks", default_value = "0")]
+    #[arg(short = 'H', long = "hawks", default_value = "0")]
     pub number_of_hawks: usize,
 
-    #[arg(short = 'd', long = "doves", default_value = "0")]
+    #[arg(short = 'D', long = "doves", default_value = "0")]
     pub number_of_doves: usize,
 
     #[arg(short = 'm', long = "mixes", default_value = "0")]
     pub number_of_mixes: usize,
 
     #[arg(short = 'M', long = "mix-hawk-probability", default_value = "0.5")]
-    pub mix_probability_of_hawk: f32,
+    pub mix_probability_of_hawk: f64,
 
     #[arg(short = 'l', long = "learners", default_value = "100")]
     pub number_of_learning: usize,
